Small cleanup and changes for function calling utility
